+++ core/src/banking_stage.rs
@@ -757,8 +757,9 @@ mod tests {
         solana_perf::packet::to_packet_batches,
         solana_poh::{
             poh_record_error::PohRecordError,
-            poh_recorder::{create_test_recorder, Record, RecordTransactionsSummary},
+            poh_recorder::{create_test_recorder, RecordTransactionsSummary},
             poh_service::PohService,
+            record::Record,
         },
         solana_runtime::{bank::Bank, genesis_utils::bootstrap_validator_stake_lamports},
         solana_runtime_transaction::runtime_transaction::RuntimeTransaction,
