+++ poh/src/poh_recorder.rs
@@ -15,6 +15,7 @@ use {
         leader_bank_notifier::LeaderBankNotifier,
         poh_record_error::{PohRecordError, Result},
         poh_service::PohService,
+        record::Record,
         working_bank_entry::WorkingBankEntry,
     },
     crossbeam_channel::{
@@ -60,32 +61,6 @@ impl BankStart {
             &self.bank_creation_time,
             self.working_bank.ns_per_slot,
         )
-    }
-}
-
-// Sends the Result of the record operation, including the index in the slot of the first
-// transaction, if being tracked by WorkingBank
-type RecordResultSender = Sender<Result<Option<usize>>>;
-
-pub struct Record {
-    pub mixin: Hash,
-    pub transactions: Vec<VersionedTransaction>,
-    pub slot: Slot,
-    pub sender: RecordResultSender,
-}
-impl Record {
-    pub fn new(
-        mixin: Hash,
-        transactions: Vec<VersionedTransaction>,
-        slot: Slot,
-        sender: RecordResultSender,
-    ) -> Self {
-        Self {
-            mixin,
-            transactions,
-            slot,
-            sender,
-        }
     }
 }
